requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0; python_version >= '3.7'
certifi==2022.12.7; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.7
django-otp==1.1.4
idna==3.4; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pycparser==2.21
pyjwt==2.6.0; python_version >= '3.7'
python-google-places==1.4.2
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.0.0
social-auth-core==4.3.0; python_version >= '3.6'
sqlparse==0.4.3; python_version >= '3.5'
urllib3==1.26.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
